requirements.txt
cohere==5.11.3
fastapi==0.115.4
httpx==0.27.2
peewee==3.17.7
pytest==8.3.3
python-dotenv==1.0.1
uvicorn==0.32.0